Parse diff header after line parse
- Move diff header parse after line parse

Only when line info is found should the header parse stop.
'parsing_diff_header' is set false in the line parse code block. Any
evaluation of that variable must happen after the line parse code block.

- In line parse abort if diff header was not properly parsed

When line info is found the header is finished so determine whether
or not it was parsed properly. Currently the only header info needed is
a path and 'found_path' is evaluated to determine whether or not it was
found.