Fix line number extraction
- Fix the line number regex to take in the whole hunk line.

The line number regex is used by gensub() and is meant to only return
the line number but to do that a wildcard must be used to take in the
rest of the hunk line for removal.

Prior to this change there was no wildcard at the end of the expression
so the extracted line number could be incorrect as a string. However
because the statement (line = line + 0;) explicitly converts that string
to a number, and because of Awk's type conversion behavior, it ended up
correct from then on as either type even though the regex was incorrect.