Add option to show binary files
@show_binary [0,1] default: ( show_header ? 0 : 1 )
Show a binary file that differs in an empty format. [path:]<:>

For example, if show_header=0 and binary file a.exe differs it will be
shown in an "empty format":
a.exe::
abc.c: :-  asdf
abc.c:1:+       test

Prior to this change there was no notice that binary files differed
unless the header was shown. That notice is still shown in that
circumstance regardless of this option's setting:
Binary files <old> and <new> differ

- (reset_header_variables) New. Move the reset of all header variables
from init() to this function so they can be reset easily at other times.

- Store the start line of the diff header in 'diff'.