Change @allow_colons_in_path default value to !@show_path
ie @show_path is on then @allow_colons_in_path is off and vice versa.

Prior to this change @allow_colons_in_path defaulted to off always, but
there's no reason to stop colons in the path if the path isn't shown so
this is a better default.

For example here's a git diff command that prior to this change would
have caused showlinenum to abort by default, but now it's processed:

git diff HEAD~1:./foo/bar HEAD:./foo/baz | showlinenum.awk